Move direct socket access into a layer of abstraction
